Separate ktranscript plugin test into its own autotest
The plugin based test of KTranscript performs all tests with a single
instance of KTranscriptImp because the plugin uses Q_GLOBAL_STATIC
to create and access that instance.

The non-plugin test creates a new instance for every test.

Currently all scripting tests are runnable in both situations, but the
non-plugin test allows for tests that need a "clean slate".

REVIEW: 115936