Remove use of KI18N_MODULE_DIR.
This variable (and the corresponding if/else in KF5I18NMacros) is
unnecessary, as the scripts needed by KF5I18NMacros.cmake are always in
the same directory as KF5I18NMacros.cmake.

We keep the variable, albeit set in a simpler way, in case anything else
is using it (by the CMake naming convention, a variable with no leading
underscore is public).

REVIEW: 122754