Move the i18n context from KDeclarative
The only way to use i18n*() so far in QML is by depending on all
KDeclarative. This patch moves the necessary bits there so it can be
adopted by an application or framework just by offering the needed bits
within KI18n.
This is done by offering an object with methods that map to the i18n*() C++
counter-parts.
Furthermore, this improves on the version in KDeclarative by letting KI18n
do the conversion from value into string that used to be done by the QML
engine before.

REVIEW: 126087