libgettext: Fix potential use-after-free using non-g++ compilers.
Coverity noted (in CID 1354239 and 1354240) that the value of a pointer
is used in a comparison after the pointer is freed (when being compiled
by non-g++ or g++ in strict ANSI mode, where variable-length arrays are
not supported).

This is technically undefined behavior and can only lead to problems
with optimizing compilers - instead do the test while the pointer is
still valid and use the test result later.

This is technically a GNU gettext bug, which has been reported upstream
(bug 47847).

REVIEW:127834