Move .po and .ts files look-up to build-time
Summary:
As specified in cmake's documentation, file(GLOB) is not recommended for looking
up sources because it requires cmake to be regenerated to be able to process
what needs to be executed.
This patch moves this process into the compilation time where we'll combine
the look-up and the generation.

Test Plan:
Tested on a project with only po files, will need testing on projects
that do include scripts.

Reviewers: #frameworks, sitter, ltoscano, ilic

Reviewed By: sitter

Subscribers: aacid

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D5167