API dox: add note about calling setApplicationDomain after QApp creation
Summary:
The current implementation of KLocalizedString::setApplicationDomain()
will trigger the creation of the KLocalizedStringPrivateStatics instance and
thus also the calculation of the locale languanges from the env vars
as set at that time, and QLocale::system() on some platforms.
So the claim that "This call can be made at any place in the code" does not
hold, instead API consumers should be instructed about when to call
the method ideally.

Reviewers: #frameworks, ilic, ltoscano, dfaure

Reviewed By: dfaure

Subscribers: dfaure

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D8351