do not treat ts-pmap-compile as exectuable
Summary:
- strip shebang
- make -x

the script is run through add_custom_command with the explicit python
exeutable found by FindPythonInterp (which may be either python
or python3) and the script is compatible with either.
To prevent confusion and prevent linting tech from tripping over the fact
that the script references the python2 interpreter (which EOLs in 2020)
we'll make sure not to make it look like this script should be run
without manually running it through the interpreter.

Reviewers: ilic

Reviewed By: ilic

Subscribers: #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D9295