Fix finding libintl when "cross"-compiling native Yocto packages
Summary:
In that case the libintl code we look for here is part of the host glibc,
and as such not in the host prefix that the CMake toolchain file searches
in. We can address this by handling the case correctly that both the
include and library paths are empty but the symbol is found nevertheless.

Reviewers: #frameworks, ilic

Reviewed By: ilic

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D9585