[API dox] New UI marker @item:valuesuffix
Summary:
Allows to mark strings which are used for concatenating them directly to a
value string, to result in some quantity string.

As e.g. useful with Qt's QSpinBox and QCC1 SpinBox, which have a suffix
string property which then is displayed concatenated to the currently
selected value.

Reviewers: #localization, ilic

Reviewed By: ilic

Subscribers: aacid, #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D12353