Fix KCatalog::translate when translation is same as original text
Summary:
We need to be compare the pointer passed to gettext with the one we get
back.

If it is the same it means no translation was found, if it
is different then it means the translation was found.

Test Plan: Test now succeeds, without patch it fails

Reviewers: ilic

Reviewed By: ilic

Subscribers: ilic, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D14473