Also add quotes around rich text <filename> tagged text
Quotes are added around the plain text version. The rich text version
instead gets a monospaced style, but in some cases the quotes are more
helpful--for example when the text begins or ends with a space. So let's
add quotes around the rich text version, too.

This is an alternative to https://invent.kde.org/frameworks/kio/-/merge_requests/15