Fix possible preprocessor race condition with i18n defines
For library that define a TRANSLATION_DOMAIN and include both the
klocalizedstring.h and the klocalizedcontext.h headers. If these
headers are not included in the stated order, then the TRANSLATION_DOMAIN
specific defines (from klocalizedstring.h) modify the method names
in klocalizedcontext.h and lead to preprocessor errors.

This patch enforces the klocalizedtext.h header to be always parsed
before the klocalizedstring.h header.