Load Qt translations even if some catalogs are missing
If loading the qt_ meta catalog fails (which may happen if loading fails
for one of the four catalogs it depends on for most locales), then
try loading the four catalogs separately. This is useful for AppImages
and other packaging formats which may include only the catalogs of
those Qt modules that are actually used.