Work around QStandardPaths not working for unit tests on Windows
QStandardPaths only provides the expected results for properly deployed
executables, which isn't the case for unit tests. Therefore any test
relying on finding message catalogs etc will fail.

As a workaround for that we explicitly add the locale data directory of
the install location to the search paths. In order to also work with Craft,
we can't use CMAKE_INSTALL_PREFIX for this (that would point to a staging
location), but fortunately we get the install location from FindIsoCodes
already.