KuitSetup: fix setting classification of tags
Old code in KuitSetupPrivate::setTagPattern(..) &
KuitSetupPrivate::setTagClass(...) had inverted logic for detecting if the
tag was already defined before. As result all built-in tags that should
have been of class StructTag were also of class PhraseTag.

That being fixed uncovered an issue in
KuitFormatterPrivate::toVisualText(...), which before when the xml string
was started to be processed and the logic for finding the enclosing element
was reached with the root element <__kuit_internal_top__>,
the context so far was empty, thus also the element stack openEls.
So no hit in the loop was found and parseOpenEl() was called with
the default-constructed enclosingOel. That by accident just happened
to work, but no longer does with "kuit_internal_top" now properly set
to be a StructTag. Doing a special-case handling of the root element
instead solves this.