Directly load catalogs from Android assets
This avoids having to copy catalogs from assets into the file system cache
first. That's inefficient, and worse, we are actually never updating those
copies (and that is non-trivial as assets have no file metadata to notice
when that might be necessary).

The downside of this is the need for fmemopen(), which only exists starting
at API level 23. We therefore can't use that directly and have to check
for it at runtime using dlsym() as well as keeping the old path with all
its shortcomings around as a fallback.