Add iso-codes cache generation tool
This adds a simple command line tool for pre-generating the iso-codes
caches currently created at runtime on first use.

The goal here is to eventually make use of this in bundled builds
(APKs, etc) by baking in the cache files into a qrc and not needing
the iso-codes JSON files at runtime at all anymore. As the iso-codes
data cannot be updated independently in those setups anyway we can
avoid the overhead for supporting that as well.