Load translations for application-specific language also on Win and Mac
On Windows and Apple OSs, we cannot use QLocale::system() because on
those OSs Qt ignores LANGUAGE (which is set by kxmlgui to the
application-specific languages configured by the user). We need to
check LANGUAGE ourselves.

CCBUG: 464694

This MR replaces kxmlgui!145. It fixes the bug for the translations provided by Qt. To fix the bug also for translations coming from .qm files in KF libraries a similar change in ecm is needed.

I have tested this MR (for KF5) with Kate on Windows.