KF5I18nMacros.cmake.in - don't look for python[2,3] on Windows
Windows doesn't have versioned python executables.
However, pythonX could be found in an msys64 or a cygwin path;
but the linux-y pythons won't work here since they think
things like path delimiters are ':' instead of the Windows ';'.

So force "python" as the executable on Windows and expect
the user to be setting their %PATH% properly.