Add helper for spinbox i18n
This helper enables i18n for localising the Q*SpinBox'es. Normally for QtWidgets applications, it is only possible to add a prefix or a suffix, which does not play well with languages with different word or operator orders.

QtQuick does not have this limitation as far as it seems, but there are a lot of applications that would benefit from this.

Credit goes to Lukas Sommer for this neat implementation.