Special-case the language fallback for country-less English
For any other language we need a fallback chain from country-specific to
the country-less variants, only for English we don't have a country-less
variant but the en_US source code.

Without this "en" translations aren't found and we continue trying other
languages before eventually getting to "en_US". This results in unexpected
results when using a country-specific variant of English that doesn't have
translations and at least one secondary non-English language.

This is unusual though possible on Linux, but common on Android. There
language and region can be configured independently (and thus it's easy
to get to an English country-specific variant without translations) and
multiple languages are common as those not only influence translations but
also input handling.